Changed noise from float64 to float32
Former-commit-id: d2df019e7e3b93fb7fdccf5fa1494d455825f1c6
Former-commit-id: 055d4955b9e5810d5bf20b194e7ec1f3d429032d

+++ modules/transformations/audio.py
@@ -41,7 +41,7 @@ class Noise:
         if random.random() < self.p:
             RMS_s = np.sqrt(np.mean(audio**2))
             RMS_n = np.sqrt(RMS_s**2 / (pow(10, self.snr / 20)))
-            noise = np.random.normal(0, RMS_n, audio.shape[0])
+            noise = np.random.normal(0, RMS_n, audio.shape[0]).astype('float32')
             audio = audio + noise
             audio = np.clip(audio, -1, 1)
         return audio
